Refactor: Implement user-specific CRUD operations for expenses
- Ensure that users can only perform CRUD operations on their own expenses
- Implement CRUD operations for expenses by both regular users and admins
- Remove unnecessary and commented-out code

+++ controllers/users.controller.js
@@ -3,6 +3,16 @@ const crypto = require('crypto-js');
 
 
 async function CreateNewUser(data) {
+
+    //Check if username already exists
+    const userNameAlreadyExists = await UsersModel.findOne({
+        where:{
+            username: data.username,
+        } 
+    });
+    if(userNameAlreadyExists){
+        throw new Error("Bad Request: Username already exists")
+    }
     //Creating user in DB
     // Hash the password before saving it
     const hashedPassword = crypto.SHA256(data.password.toString()).toString();
@@ -11,20 +21,41 @@ async function CreateNewUser(data) {
     return createUser;
 }
 
-async function UpdateUser(userId, newData) {
-    //Update User
+async function UpdateUser(userId, newData, loggedInUserData) {
+    //Check if user exists
     const getUser = await UsersModel.findByPk(userId);
 
     if (!getUser) {
         throw new Error("User not found")
     }
 
+    //Check for user to update only his profile
+    if (loggedInUserData.usertype !== "admin") {
+        if (getUser.username !== loggedInUserData.username) {
+            throw new Error("Forbidden - You do not have permission to perform this operation")
+        }
+    }
+
+    //Update User
     getUser.set(newData)
     await getUser.save();
     return getUser
 }
 
-async function DeleteUser(id) {
+async function DeleteUser(id, loggedInUserData) {
+
+    //check if user exists
+    const getUser = await UsersModel.findByPk(userId);
+
+    if (!getUser) {
+        throw new Error("User not found")
+    }
+
+    //Only admin can delete users
+    if (loggedInUserData.usertype !== "admin") {
+            throw new Error("Forbidden - You do not have permission to perform this operation")
+    }
+
     //Delete User
     const deletedUser = await UsersModel.destroy({
         where: { id },
